// webclient/src/api/generated/dtos/tickerArticles.gen.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Sentiment Search API
 * Three-stage pipeline: (1) stream stock info, (2) stream articles per ticker, (3) stream per-article sentiment scores.

 * OpenAPI spec version: 2.0.0
 */
import type { TickerArticlesSourcesItem } from './tickerArticlesSourcesItem.gen';

export interface TickerArticles {
  ticker: string;
  sources: TickerArticlesSourcesItem[];
}